Day 17 with some Notes
On the rising edge of valid_i / req_i, generate a random value using day_7 LFSR module.
Rising edge of valid_i / req_i is detected using day_3 module.
This is to generate the ready_o after a random delay.
After random value is generated, it counts by 1 every positive clock cycle.
As soon as the count reaches to 0, the ready_o is made logic 1 and data is written into memory if write s/g is high.
It reads from the memory when read s/g is valid.